+++ tests/testthat/test-cff_write.R
@@ -3,13 +3,14 @@ test_that("Write basic", {
     system.file("examples/DESCRIPTION_basic", package = "cffr")
   desc_file <- cff_create(desc_file)
   expect_s3_class(desc_file, "cff")
-  tmp <- tempfile(fileext = ".cff")
+  tmp <- file.path(tempdir(), "basic.cff")
   expect_message(cff_write(desc_file, outfile = tmp, validate = FALSE))
   expect_silent(cff_write(desc_file, outfile = tmp, verbose = FALSE))
   expect_true(file_exist_abort(tmp))
 
   # Validate from file
   expect_true(cff_validate(tmp, verbose = FALSE))
+  expect_snapshot_file(tmp)
 
   file.remove(tmp)
   expect_false(file_exist_abort(tmp))
@@ -20,7 +21,7 @@ test_that("Write to a non-existing folder", {
     system.file("examples/DESCRIPTION_basic", package = "cffr")
   desc_file <- cff_create(desc_file)
   expect_s3_class(desc_file, "cff")
-  tmp <- tempfile("/test_new_folder/recursive/", fileext = ".cff")
+  tmp <- file.path(tempdir(), "/test_new_folder/recursive/new_folder.cff")
   cff_write(
     desc_file,
     outfile = tmp,
@@ -35,6 +36,7 @@ test_that("Write to a non-existing folder", {
 
   # Validate from file
   expect_true(cff_validate(tmp, verbose = FALSE))
+  expect_snapshot_file(tmp)
 
   file.remove(tmp)
   expect_false(file_exist_abort(tmp))
@@ -45,7 +47,7 @@ test_that("Write no encoding", {
     system.file("examples/DESCRIPTION_no_encoding", package = "cffr")
   desc_file <- cff_create(desc_file)
   expect_s3_class(desc_file, "cff")
-  tmp <- tempfile(fileext = ".cff")
+  tmp <- file.path(tempdir(), "noencoding.cff")
   cff_write(
     desc_file,
     outfile = tmp,
@@ -57,6 +59,7 @@ test_that("Write no encoding", {
 
   # Validate from file
   expect_true(cff_validate(tmp, verbose = FALSE))
+  expect_snapshot_file(tmp)
 
   file.remove(tmp)
   expect_false(file_exist_abort(tmp))
@@ -67,7 +70,7 @@ test_that("Add new keys", {
     system.file("examples/DESCRIPTION_basic", package = "cffr")
   desc_file <- cff_create(desc_file)
   expect_s3_class(desc_file, "cff")
-  tmp <- tempfile(fileext = ".cff")
+  tmp <- file.path(tempdir(), "newkeys.cff")
 
 
   # Add additional keys
@@ -96,6 +99,7 @@ test_that("Add new keys", {
 
   # Validate from file
   expect_true(cff_validate(tmp, verbose = FALSE))
+  expect_snapshot_file(tmp)
 
   file.remove(tmp)
   expect_false(file_exist_abort(tmp))
@@ -126,7 +130,7 @@ test_that("Append keys", {
     )
   )
 
-  tmp <- tempfile(fileext = ".cff")
+  tmp <- file.path(tempdir(), "appendkeys.cff")
   s <- cff_write(
     desc_file,
     outfile = tmp,
@@ -137,6 +141,7 @@ test_that("Append keys", {
   expect_snapshot(s)
 
   expect_true(cff_validate(tmp, verbose = FALSE))
+  expect_snapshot_file(tmp)
 
   file.remove(tmp)
   expect_false(file_exist_abort(tmp))
@@ -144,14 +149,46 @@ test_that("Append keys", {
 
 test_that("Fix extension of the file", {
   cffobj <- cff()
-
-  tmp <- tempfile()
+  cffobj <- cff_modify(cffobj, authors = as_cff_person("Diego Pérez"))
+  tmp <- file.path(tempdir(), "fix_extension")
   expect_silent(cff_write(cffobj, tmp, verbose = FALSE))
 
   expect_false(file_exist_abort(tmp))
   expect_true(file_exist_abort(paste0(tmp, ".cff")))
   expect_true(cff_validate(paste0(tmp, ".cff"), verbose = FALSE))
+  expect_snapshot_file(paste0(tmp, ".cff"))
 
   file.remove(paste0(tmp, ".cff"))
   expect_false(file_exist_abort(paste0(tmp, ".cff")))
+})
+
+test_that("test encoding utf8", {
+  cffobj <- cff()
+  cffobj <- cff_modify(cffobj, authors = as_cff_person("Diego Pérez"))
+  tmp <- file.path(tempdir(), "utf8.cff")
+  expect_silent(cff_write(cffobj, tmp, verbose = FALSE))
+
+  expect_true(file_exist_abort(tmp))
+  expect_true(cff_validate(tmp, verbose = FALSE))
+  expect_snapshot_file(tmp)
+
+  file.remove(tmp)
+  expect_false(file_exist_abort(tmp))
+})
+
+test_that("test encoding others", {
+  cffobj <- cff()
+  cffobj <- cff_modify(cffobj, authors = as_cff_person("Diego Pérez"))
+  tmp <- file.path(tempdir(), "asci_trans.cff")
+  expect_silent(cff_write(cffobj, tmp,
+    verbose = FALSE,
+    encoding = "ASCII//TRANSLIT"
+  ))
+
+  expect_true(file_exist_abort(tmp))
+  expect_true(cff_validate(tmp, verbose = FALSE))
+  expect_snapshot_file(tmp)
+
+  file.remove(tmp)
+  expect_false(file_exist_abort(tmp))
 })